Add an obscure incantation
As a workaround for https://github.com/microbit-foundation/micropython-microbit-v2/issues/83 it was suggested by the micropython team to poke the pin configuration register directly (under the nose of CODAL). Whilst this works Im not convinced its a 'good idea' as a change to the underlying system could render this 'fix' broken. 
It has been tested on V1 and V2 and doesn't appear to have any bad side effects.